# docs/v1/openapi.yml
openapi: 3.1.0
info:
  title: Ocha API
  version: 1.0.0
  description: |
    API REST pour la gestion de matcha shops Ocha, incluant la gestion des utilisateurs, produits, magasins et commandes.
    
    ## Fonctionnalités
    - **Gestion de l'authentification** : Inscription et connexion des utilisateurs avec JWT.
    - **Gestion des produits** : Création, mise à jour, suppression et récupération des produits.
    - **Gestion des magasins** : Création, mise à jour, suppression et récupération des magasins.
    - **Gestion des commandes** : Création de commandes par les utilisateurs, mise à jour du statut
    - **Statistiques des commandes** : Récupération des statistiques des commandes par utilisateur.

    ## Authentification
    - **Inscription** : les utilisateurs peuvent s'inscrire en fournissant une adresse e-mail, un mot de passe et un nom d'affichage, eventuellement un numéro de téléphone.
    - **Connexion** : les utilisateurs peuvent se connecter avec leur adresse e-mail et mot de passe pour obtenir un token JWT.
    - **Sécurité** : les endpoints protégés nécessitent un token JWT dans l'en-tête `Authorization Bearer <token>`.
    - **Rôles** : les utilisateurs ont des rôles (`user`, `admin`) qui contrôlent l'accès aux fonctionnalités.
    - **Hashage des mots de passe** : les mots de passe sont stockés de manière sécurisée en utilisant bcrypt.

    ## Code de status de réponse HTTP pouvant être rencontrés
    - `200 OK` : Requête réussie (ressource retournée).
...

...

...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Récupérer la liste des produits
      description: |
        Retourne une liste paginée des produits disponibles.
        Les produits incluent les tailles disponibles et les prix supplémentaires
        par taille.
      operationId: getProducts
      parameters:
        - in: query
          name: page
          required: false
          description: Numéro de page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          description: Nombre de produits par page
          schema:
            type: integer
            minimum: 1
            default: 10
        - in: query
          name: isActive
          required: false
          description: Filtrer les produits actifs/inactifs
          schema:
            type: boolean
      responses:
        "200":
          description: Liste des produits
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "List of products"
                  page:
                    type: integer
                    example: 1
                  totalPages:
                    type: integer
                    example: 10
                  totalProducts:
                    type: integer
                    example: 10
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          format: mongo-id
                        slug:
                          type: string
                          example: "classic-matcha-latte"
                        name:
                          type: string
                          example: "Classic Matcha Latte"
                        category:
                          type: string
                          example: "Matcha"
                        description:
                          type: string
                        basePriceCHF:
                          type: number
                          example: 5.9
                        isActive:
                          type: boolean
                          example: true
                        image:
                          type: string
                          format: uri
                        size:
                          type: array
                          items:
                            type: string
                            enum: [S, M, L]
                        extra_chf:
                          type: object
                          additionalProperties:
                            type: number
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                        __v:
                          type: integer
              example:
                message: "List of products"
                page: 1
                totalPages: 10
                totalProducts: 10
                products:
                  - _id: "695fb951bf4a414b74147b9e"
                    slug: "classic-matcha-latte"
                    name: "Classic Matcha Latte"
                    category: "Matcha"
                    description: "Matcha latte classique, équilibré et crémeux."
                    basePriceCHF: 5.9
                    isActive: true
                    image: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/classic-matcha-latte.jpg"
                    size: ["S", "M", "L"]
                    extra_chf:
                      S: 0
                      M: 2
                      L: 3
                    createdAt: "2026-01-08T14:04:01.597Z"
                    updatedAt: "2026-01-08T14:04:01.597Z"
                    __v: 0
        "500":
          description: Internal server error

    post:
      tags:
        - Products
      summary: Créer un produit
      description: |
        Crée un nouveau produit Ocha.
        Endpoint réservé aux administrateurs.
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - slug
                - name
                - category
                - description
                - basePriceCHF
                - image
                - size
                - extra_chf
              properties:
                slug:
                  type: string
                  example: "white-matcha-latte"
                name:
                  type: string
                  example: "White Matcha Latte"
                category:
                  type: string
                  example: "Boissons"
                description:
                  type: string
                  example: "Delicious strawberry matcha latte"
                basePriceCHF:
                  type: number
                  example: 5
                image:
                  type: string
                  format: uri
                  example: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/mixed-berries-matcha-latte.jpg"
                isActive:
                  type: boolean
                  default: true
                size:
                  type: array
                  items:
                    type: string
                    enum: [S, M, L]
                extra_chf:
                  type: object
                  additionalProperties:
                    type: number
                  example:
                    S: 0
                    M: 2
                    L: 3
            example:
              slug: "white-matcha-latte"
              name: "White Matcha Latte"
              category: "Boissons"
              description: "Delicious strawberry matcha latte"
              basePriceCHF: 5
              image: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/mixed-berries-matcha-latte.jpg"
              isActive: true
              size: ["S", "M", "L"]
              extra_chf:
                S: 0
                M: 2
                L: 3
      responses:
        "201":
          description: Produit créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product created"
                  product:
                    type: object
                    properties:
                      _id:
                        type: string
                        format: mongo-id
                        example: "6962eb966aaa44672169d87e"
                      slug:
                        type: string
                      name:
                        type: string
                      category:
                        type: string
                      description:
                        type: string
                      basePriceCHF:
                        type: number
                      isActive:
                        type: boolean
                      image:
                        type: string
                        format: uri
                      size:
                        type: array
                        items:
                          type: string
                      extra_chf:
                        type: object
                        additionalProperties:
                          type: number
                      createdAt:
                        type: string
                        format: date-time
                        example: "2026-01-11T00:15:18.543Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2026-01-11T00:15:18.543Z"
                      __v:
                        type: integer
                        example: 0
              example:
                message: "Product created"
                product:
                  slug: "white-matcha-latte"
                  name: "White Matcha Latte"
                  category: "Boissons"
                  description: "Delicious strawberry matcha latte"
                  basePriceCHF: 5
                  isActive: true
                  image: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/mixed-berries-matcha-latte.jpg"
                  size: ["S", "M", "L"]
                  extra_chf:
                    S: 0
                    M: 2
                    L: 3
                  _id: "6962eb966aaa44672169d87e"
                  createdAt: "2026-01-11T00:15:18.543Z"
                  updatedAt: "2026-01-11T00:15:18.543Z"
                  __v: 0
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict
        "422":
          description: Validation error
        "500":
          description: Internal server error
  /products/{id}:
    get:
      tags:
        - Products
      summary: Récupérer un produit par ID
      description: |
        Retourne les détails complets d’un produit Ocha à partir de son identifiant.
      operationId: getProductById
      parameters:
        - in: path
          name: id
          required: true
          description: Identifiant MongoDB du produit
          schema:
            type: string
            format: mongo-id
      responses:
        "200":
          description: Produit trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product found"
                  product:
                    type: object
                    properties:
                      _id:
                        type: string
                        format: mongo-id
                        example: "695fb951bf4a414b74147b9e"
                      slug:
                        type: string
                        example: "classic-matcha-latte"
                      name:
                        type: string
                        example: "Classic Matcha Latte"
                      category:
                        type: string
                        example: "Matcha"
                      description:
                        type: string
                        example: "Matcha latte classique, équilibré et crémeux."
                      basePriceCHF:
                        type: number
                        example: 5.9
                      isActive:
                        type: boolean
                        example: true
                      image:
                        type: string
                        format: uri
                        example: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/classic-matcha-latte.jpg"
                      size:
                        type: array
                        items:
                          type: string
                          enum: [S, M, L]
                      extra_chf:
                        type: object
                        additionalProperties:
                          type: number
                        example:
                          S: 0
                          M: 2
                          L: 3
                      createdAt:
                        type: string
                        format: date-time
                        example: "2026-01-08T14:04:01.597Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2026-01-08T14:04:01.597Z"
                      __v:
                        type: integer
                        example: 0
              example:
                message: "Product found"
                product:
                  _id: "695fb951bf4a414b74147b9e"
                  slug: "classic-matcha-latte"
                  name: "Classic Matcha Latte"
                  category: "Matcha"
                  description: "Matcha latte classique, équilibré et crémeux."
                  basePriceCHF: 5.9
                  isActive: true
                  image: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/classic-matcha-latte.jpg"
                  size: ["S", "M", "L"]
                  extra_chf:
                    S: 0
                    M: 2
                    L: 3
                  createdAt: "2026-01-08T14:04:01.597Z"
                  updatedAt: "2026-01-08T14:04:01.597Z"
                  __v: 0
        "400":
          description: Bad request
        "404":
          description: Product not found
        "500":
          description: Internal server error

    patch:
      tags:
        - Products
      summary: Mettre à jour partiellement un produit
      description: |
        Met à jour un ou plusieurs champs d’un produit existant.
        Endpoint réservé aux administrateurs.
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du produit
          schema:
            type: string
            format: mongo-id
            example: "6962eb966aaa44672169d87e"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                slug:
                  type: string
                name:
                  type: string
                category:
                  type: string
                description:
                  type: string
                basePriceCHF:
                  type: number
                isActive:
                  type: boolean
                image:
                  type: string
                  format: uri
                size:
                  type: array
                  items:
                    type: string
                    enum: [S, M, L]
                extra_chf:
                  type: object
                  additionalProperties:
                    type: number
            example:
              extra_chf:
                S: 0
                M: 3
                L: 6
      responses:
        "200":
          description: Produit mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product updated"
                  product:
                    type: object
                    properties:
                      _id:
                        type: string
                        format: mongo-id
                      slug:
                        type: string
                      name:
                        type: string
                      category:
                        type: string
                      description:
                        type: string
                      basePriceCHF:
                        type: number
                      isActive:
                        type: boolean
                      image:
                        type: string
                        format: uri
                      size:
                        type: array
                        items:
                          type: string
                      extra_chf:
                        type: object
                        additionalProperties:
                          type: number
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                      __v:
                        type: integer
              example:
                message: "Product updated"
                product:
                  _id: "6962eb966aaa44672169d87e"
                  slug: "white-matcha-latte"
                  name: "White Matcha Latte"
                  category: "Boissons"
                  description: "Delicious strawberry matcha latte"
                  basePriceCHF: 5
                  isActive: true
                  image: "https://res.cloudinary.com/dabosy2w2/image/upload/v1765369784/mixed-berries-matcha-latte.jpg"
                  size: ["S", "M", "L"]
                  extra_chf:
                    S: 0
                    M: 3
                    L: 6
                  createdAt: "2026-01-11T00:15:18.543Z"
                  updatedAt: "2026-01-11T00:18:19.260Z"
                  __v: 0
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Product not found
        "422":
          description: Validation error
        "500":
          description: Internal server error

    delete:
      tags:
        - Products
      summary: Supprime un produit
      description: Supprime un produit existant par son identifiant.
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Product deleted successfully
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Product not found
        "500":
          description: Internal server error

    put:
      tags:
        - Products
      summary: Mettre à jour l’image d’un produit
      description: |
        Remplace l’image associée à un produit existant.
        Retourne l’URL finale de l’image stockée (Cloudinary).
      operationId: uploadProductImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du produit
          schema:
            type: string
            format: mongo-id
            example: "6962eb966aaa44672169d87e"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - image
              properties:
                image:
                  type: string
                  format: uri
                  description: URL de l’image à associer au produit
            example:
              image: "https://res.cloudinary.com/dabosy2w2/image/upload/v1768091151/products/white-matcha-latte.jpg"
      responses:
        "200":
          description: Image mise à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Image uploaded successfully"
                  imageUrl:
                    type: string
                    format: uri
              example:
                message: "Image uploaded successfully"
                imageUrl: "https://res.cloudinary.com/dabosy2w2/image/upload/v1768091192/products/white-matcha-latte.jpg"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Product not found
        "422":
          description: Validation error
        "500":
          description: Internal server error

  # ---------------------------------------
  # Stores
  # ---------------------------------------
...